Please enter the commit message for your changes. Lines starting implement rating breakdown bars
